+++ examples/demo_redis.py
@@ -5,7 +5,7 @@ Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
 from cryptofeed import FeedHandler
-from cryptofeed.backends.redis import BookDeltaStream, BookStream, FundingRedis, OpenInterestRedis, TradeRedis, CandlesStream
+from cryptofeed.backends.redis import BookDeltaStream, BookStream, CandlesRedis, FundingRedis, OpenInterestRedis, TradeRedis
 from cryptofeed.defines import CANDLES, FUNDING, L2_BOOK, OPEN_INTEREST, TRADES, BOOK_DELTA
 from cryptofeed.exchanges import Bitfinex, Bitmex, Coinbase, Gemini, Binance
 
@@ -17,7 +17,7 @@ def main():
     f.add_feed(Coinbase(channels=[TRADES], symbols=['BTC-USD'], callbacks={TRADES: TradeRedis()}))
     f.add_feed(Coinbase(channels=[L2_BOOK], symbols=['BTC-USD'], callbacks={BOOK_DELTA: BookDeltaStream(), L2_BOOK: BookStream()}))
     f.add_feed(Gemini(symbols=['BTC-USD'], callbacks={TRADES: TradeRedis()}))
-    f.add_feed(Binance(symbols=['BTC-USDT'], channels=[CANDLES], callbacks={CANDLES: CandlesStream()}))
+    f.add_feed(Binance(candle_closed_only=True, symbols=['BTC-USDT'], channels=[CANDLES], callbacks={CANDLES: CandlesRedis(score_key='start')}))
 
     f.run()